Rename RootNavigationCtl
